tests pass, time to clean up
