--- ocflv1/object.go
@@ -1,232 +1,0 @@
-package ocflv1
-
-import (
-	"cmp"
-	"context"
-	"errors"
-	"fmt"
-	"io"
-	"io/fs"
-	"path"
-	"strings"
-	"time"
-
-	"github.com/srerickson/ocfl-go"
-	"golang.org/x/exp/slices"
-)
-
-var (
-	ErrOCFLVersion      = errors.New("unsupported OCFL version")
-	ErrObjRootStructure = errors.New("object includes invalid files or directories")
-)
-
-// ReadObject implements ocfl.ReadObject for OCFL v1.x objects
-type ReadObject struct {
-	fs   ocfl.FS
-	path string
-	inv  *Inventory
-}
-
-func (o *ReadObject) FS() ocfl.FS { return o.fs }
-
-func (o *ReadObject) Inventory() ocfl.Inventory {
-	if o.inv == nil {
-		return nil
-	}
-	return &readInventory{raw: *o.inv}
-}
-
-func (o *ReadObject) VersionFS(ctx context.Context, i int) fs.FS {
-	ver := o.inv.Version(i)
-	if ver == nil {
-		return nil
-	}
-	// FIXME: This is a hack to make versionFS replicates the filemode of
-	// the undering FS. Open a random content file to get the file mode used by
-	// the underlying FS.
-	regfileType := fs.FileMode(0)
-	for _, paths := range o.inv.Manifest {
-		if len(paths) < 1 {
-			break
-		}
-		f, err := o.fs.OpenFile(ctx, path.Join(o.path, paths[0]))
-		if err != nil {
-			return nil
-		}
-		defer f.Close()
-		info, err := f.Stat()
-		if err != nil {
-			return nil
-		}
-		regfileType = info.Mode().Type()
-		break
-	}
-	return &versionFS{
-		ctx:     ctx,
-		obj:     o,
-		paths:   ver.State.PathMap(),
-		created: ver.Created,
-		regMode: regfileType,
-	}
-}
-
-func (o *ReadObject) Path() string { return o.path }
-
-type versionFS struct {
-	ctx     context.Context
-	obj     *ReadObject
-	paths   ocfl.PathMap
-	created time.Time
-	regMode fs.FileMode
-}
-
-func (vfs *versionFS) Open(logical string) (fs.File, error) {
-	if !fs.ValidPath(logical) {
-		return nil, &fs.PathError{
-			Err:  fs.ErrInvalid,
-			Op:   "open",
-			Path: logical,
-		}
-	}
-	if logical == "." {
-		return vfs.openDir(".")
-	}
-	digest := vfs.paths[logical]
-	if digest == "" {
-		// name doesn't exist in state.
-		// try opening as a directory
-		return vfs.openDir(logical)
-	}
-
-	realNames := vfs.obj.inv.Manifest[digest]
-	if len(realNames) < 1 {
-		return nil, &fs.PathError{
-			Err:  fs.ErrNotExist,
-			Op:   "open",
-			Path: logical,
-		}
-	}
-	realName := realNames[0]
-	if !fs.ValidPath(realName) {
-		return nil, &fs.PathError{
-			Err:  fs.ErrInvalid,
-			Op:   "open",
-			Path: logical,
-		}
-	}
-	f, err := vfs.obj.fs.OpenFile(vfs.ctx, path.Join(vfs.obj.path, realName))
-	if err != nil {
-		err = fmt.Errorf("opening file with logical path %q: %w", logical, err)
-		return nil, err
-	}
-	return f, nil
-}
-
-func (vfs *versionFS) openDir(dir string) (fs.File, error) {
-	prefix := dir + "/"
-	if prefix == "./" {
-		prefix = ""
-	}
-	children := map[string]*vfsDirEntry{}
-	for p := range vfs.paths {
-		if !strings.HasPrefix(p, prefix) {
-			continue
-		}
-		name, _, isdir := strings.Cut(strings.TrimPrefix(p, prefix), "/")
-		if _, exists := children[name]; exists {
-			continue
-		}
-		entry := &vfsDirEntry{
-			name:    name,
-			mode:    vfs.regMode,
-			created: vfs.created,
-			open:    func() (fs.File, error) { return vfs.Open(path.Join(dir, name)) },
-		}
-		if isdir {
-			entry.mode = entry.mode | fs.ModeDir | fs.ModeIrregular
-		}
-		children[name] = entry
-	}
-	if len(children) < 1 {
-		return nil, &fs.PathError{
-			Op:   "open",
-			Path: dir,
-			Err:  fs.ErrNotExist,
-		}
-	}
-
-	dirFile := &vfsDirFile{
-		name:    dir,
-		entries: make([]fs.DirEntry, 0, len(children)),
-	}
-	for _, entry := range children {
-		dirFile.entries = append(dirFile.entries, entry)
-	}
-	slices.SortFunc(dirFile.entries, func(a, b fs.DirEntry) int {
-		return cmp.Compare(a.Name(), b.Name())
-	})
-	return dirFile, nil
-}
-
-type vfsDirEntry struct {
-	name    string
-	created time.Time
-	mode    fs.FileMode
-	open    func() (fs.File, error)
-}
-
-var _ fs.DirEntry = (*vfsDirEntry)(nil)
-
-func (info *vfsDirEntry) Name() string      { return info.name }
-func (info *vfsDirEntry) IsDir() bool       { return info.mode.IsDir() }
-func (info *vfsDirEntry) Type() fs.FileMode { return info.mode.Type() }
-
-func (info *vfsDirEntry) Info() (fs.FileInfo, error) {
-	f, err := info.open()
-	if err != nil {
-		return nil, err
-	}
-	stat, err := f.Stat()
-	return stat, errors.Join(err, f.Close())
-}
-
-func (info *vfsDirEntry) Size() int64        { return 0 }
-func (info *vfsDirEntry) Mode() fs.FileMode  { return info.mode | fs.ModeIrregular }
-func (info *vfsDirEntry) ModTime() time.Time { return info.created }
-func (info *vfsDirEntry) Sys() any           { return nil }
-
-type vfsDirFile struct {
-	name    string
-	created time.Time
-	entries []fs.DirEntry
-	offset  int
-}
-
-var _ fs.ReadDirFile = (*vfsDirFile)(nil)
-
-func (dir *vfsDirFile) ReadDir(n int) ([]fs.DirEntry, error) {
-	if n <= 0 {
-		entries := dir.entries[dir.offset:]
-		dir.offset = len(dir.entries)
-		return entries, nil
-	}
-	if remain := len(dir.entries) - dir.offset; remain < n {
-		n = remain
-	}
-	if n <= 0 {
-		return nil, io.EOF
-	}
-	entries := dir.entries[dir.offset : dir.offset+n]
-	dir.offset += n
-	return entries, nil
-}
-
-func (dir *vfsDirFile) Close() error               { return nil }
-func (dir *vfsDirFile) IsDir() bool                { return true }
-func (dir *vfsDirFile) Mode() fs.FileMode          { return fs.ModeDir | fs.ModeIrregular }
-func (dir *vfsDirFile) ModTime() time.Time         { return dir.created }
-func (dir *vfsDirFile) Name() string               { return dir.name }
-func (dir *vfsDirFile) Read(_ []byte) (int, error) { return 0, nil }
-func (dir *vfsDirFile) Size() int64                { return 0 }
-func (dir *vfsDirFile) Stat() (fs.FileInfo, error) { return dir, nil }
-func (dir *vfsDirFile) Sys() any                   { return nil }
